Replace development base URL with production base URL
Replacing the development base URL with the production base URL to
ensure that the application uses the correct URL for production
deployment. This change ensures that the application functions correctly
in a production environment. Refactor overrides the `override_base_url`
property in the configuration files `config/defaults/settings.yml` and
`config/settings.yml`.

Fixes: #123